+++ feed/src/main/java/com/revature/feed/controller/PostController.java
@@ -1,5 +1,7 @@
 package com.revature.feed.controller;
 
+import com.auth0.jwt.interfaces.DecodedJWT;
+import com.revature.feed.config.JwtUtility;
 import com.revature.feed.models.Post;
 import com.revature.feed.models.Response;
 import com.revature.feed.services.PostService;
@@ -8,7 +10,9 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 
 @RestController("postController")
 @RequestMapping(value= "post")
@@ -17,12 +21,21 @@ public class PostController {
     private PostService postService;
 
     @Autowired
+    JwtUtility jwtUtility;
+
+    @Autowired
     public PostController(PostService postService){ this.postService = postService;}
 
 
     //Create a Post
     @PostMapping
-    public Response createPost(@RequestBody Post post){
+    public Response createPost(@RequestBody Post post, @RequestHeader Map<String, String> headers){
+        //Verify the JWT
+        DecodedJWT decoded = jwtUtility.verify(headers.get("jwt"));
+        if(decoded == null){
+            return new Response(false, "Invalid token", null);
+        }
+
         Response response;
         Post tempPost = this.postService.createPost(post);
         if(tempPost != null){
@@ -33,22 +46,38 @@ public class PostController {
         return response;
     }
 
-    //Get All posts
-    @GetMapping("feed/{pageNumber}")
-    public Response getAllPost(@PathVariable Integer pageNumber, Pageable pageable){
+    @GetMapping("fave/{pageNumber}")
+    public Response getPostFromFave(@PathVariable Integer pageNumber, @RequestHeader Map<String, String> headers){
+        //Verify the JWT
+        DecodedJWT decoded = jwtUtility.verify(headers.get("jwt"));
+        if(decoded == null){
+            return new Response(false, "Invalid token", null);
+        }
+//////////// need to modify this when rabbitmq is up and running
         Response response;
-        Page<Post> allPosts= this.postService.getAllPosts(pageNumber, pageable);
-        if(allPosts != null){
-            response = new Response(true, "here are 20 posts", allPosts);
+        Integer userId = 2;
+        Integer userId1 = 8;
+        List<Integer> exampleList = new ArrayList<>();
+        exampleList.add(userId);
+        exampleList.add(userId1);
+        List<Post> favePost = this.postService.selectPostForFav(pageNumber,exampleList);
+        if(favePost != null){
+            response = new Response(true,"Favorite list", favePost);
         }else{
-            response = new Response(false,"failed to find that page",null);
+            response = new Response(false,"You have reached the end", null);
         }
         return response;
     }
 
     //Read a post
     @GetMapping("{postId}")
-    public Response lookForAPost(@PathVariable Integer postId){
+    public Response lookForAPost(@PathVariable Integer postId, @RequestHeader Map<String, String> headers){
+        //Verify the JWT - Andrew
+        DecodedJWT decoded = jwtUtility.verify(headers.get("jwt"));
+        if(decoded == null){
+            return new Response(false, "Invalid token", null);
+        }
+
         Response response;
         Post post = this.postService.getPostById(postId);
         if(post != null){
@@ -59,10 +88,29 @@ public class PostController {
         return response;
     }
 
+    @GetMapping("comment/{postId}")
+    public Response getComment(@PathVariable Integer postId, @RequestHeader Map<String, String> headers){
+        //Verify the JWT
+        DecodedJWT decoded = jwtUtility.verify(headers.get("jwt"));
+        if(decoded == null){
+            return new Response(false, "Invalid token", null);
+        }
+        Response response;
+        List<Post> comment = this.postService.getAllParentId(postId);
+        //Need to validate info being returned
+        return response = new Response(true, "Here are the comments", comment);
+    }
+
+
     //Get Post by UserId
-    //Change to send newest to oldest
     @GetMapping("userId/{userId}")
-    public Response lookForPostByUser(@PathVariable Integer userId){
+    public Response lookForPostByUser(@PathVariable Integer userId, @RequestHeader Map<String, String> headers){
+        //Verify the JWT
+        DecodedJWT decoded = jwtUtility.verify(headers.get("jwt"));
+        if(decoded == null){
+            return new Response(false, "Invalid token", null);
+        }
+
         Response response;
         List<Post> post = this.postService.getPostByUserId(userId);
         //If statement checks size as array is always returned
@@ -76,7 +124,13 @@ public class PostController {
 
     //Update a post
     @PutMapping
-    public Response updatePost(@RequestBody Post post){
+    public Response updatePost(@RequestBody Post post, @RequestHeader Map<String, String> headers){
+        //Verify the JWT
+        DecodedJWT decoded = jwtUtility.verify(headers.get("jwt"));
+        if(decoded == null){
+            return new Response(false, "Invalid token", null);
+        }
+
         Response response;
         Post updatePost = this.postService.updatePost(post);
         if(updatePost != post){
@@ -89,7 +143,13 @@ public class PostController {
 
     //Delete a post
     @DeleteMapping("{postId}")
-    public Response deletePost(@PathVariable Integer postId){
+    public Response deletePost(@PathVariable Integer postId, @RequestHeader Map<String, String> headers){
+        //Verify the JWT
+        DecodedJWT decoded = jwtUtility.verify(headers.get("jwt"));
+        if(decoded == null){
+            return new Response(false, "Invalid token", null);
+        }
+
         Response response;
         Post deletePost = this.postService.deletePost(postId);
         if(deletePost != null){
